Moved algorithm and factory into meta category
